remove protection for follower
